staging/pkg/mod/golang.org/x/net@v0.0.0-20201110031124-69a78807bb2b/html/testdata/webkit/tests22.dat

#data
<a><b><big><em><strong><div>X</a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,33): adoption-agency-1.3
(1,33): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|         <big>
|           <em>
|             <strong>
|     <big>
|       <em>
|         <strong>
|           <div>
|             <a>
|               "X"

#data
<a><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8>A</a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|     <b>
|       <div>
|         id="1"
|         <a>
|         <div>
|           id="2"
|           <a>
|           <div>
|             id="3"
|             <a>
|             <div>
|               id="4"
|               <a>
|               <div>
|                 id="5"
|                 <a>
|                 <div>
|                   id="6"
|                   <a>
|                   <div>
|                     id="7"
|                     <a>
|                     <div>
|                       id="8"
|                       <a>
|                         "A"

#data
<a><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8><div id=9>A</a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,101): adoption-agency-1.3
(1,101): adoption-agency-1.3
(1,101): adoption-agency-1.3
(1,101): adoption-agency-1.3
(1,101): adoption-agency-1.3
(1,101): adoption-agency-1.3
(1,101): adoption-agency-1.3
(1,101): adoption-agency-1.3
(1,101): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|     <b>
|       <div>
|         id="1"
|         <a>
|         <div>
|           id="2"
|           <a>
|           <div>
|             id="3"
|             <a>
|             <div>
|               id="4"
|               <a>
|               <div>
|                 id="5"
|                 <a>
|                 <div>
|                   id="6"
|                   <a>
|                   <div>
|                     id="7"
|                     <a>
|                     <div>
|                       id="8"
|                       <a>
|                         <div>
|                           id="9"
|                           "A"

#data
<a><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8><div id=9><div id=10>A</a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,112): adoption-agency-1.3
(1,112): adoption-agency-1.3
(1,112): adoption-agency-1.3
(1,112): adoption-agency-1.3
(1,112): adoption-agency-1.3
(1,112): adoption-agency-1.3
(1,112): adoption-agency-1.3
(1,112): adoption-agency-1.3
(1,112): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|     <b>
|       <div>
|         id="1"
|         <a>
|         <div>
|           id="2"
|           <a>
|           <div>
|             id="3"
|             <a>
|             <div>
|               id="4"
|               <a>
|               <div>
|                 id="5"
|                 <a>
|                 <div>
|                   id="6"
|                   <a>
|                   <div>
|                     id="7"
|                     <a>
|                     <div>
|                       id="8"
|                       <a>
|                         <div>
|                           id="9"
|                           <div>
|                             id="10"
|                             "A"

#data
<cite><b><cite><i><cite><i><cite><i><div>X</b>TEST
#errors
(1,6): expected-doctype-but-got-start-tag
(1,46): adoption-agency-1.3
(1,50): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <cite>
|       <b>
|         <cite>
|           <i>
|             <cite>
|               <i>
|                 <cite>
|                   <i>
|       <i>
|         <i>
|           <div>
|             <b>
|               "X"
|             "TEST"
